feat: create rtk query api request
