+++ package.json
@@ -41,6 +41,7 @@
     "jsdom": "^21.1.1",
     "postcss-normalize": "^10.0.1",
     "react-hook-form": "^7.43.8",
+    "react-paginate": "^8.1.5",
     "sass": "^1.58.0",
     "typescript": "^4.9.3",
     "vite": "^4.1.0",
