+++ src/App.tsx
@@ -5,7 +5,7 @@ import About from "./Pages/About/About";
 import Page404 from "./Pages/Page404/Page404";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Employees from "./Pages/Employees/Employees";
-import ImagesGallery from "./Pages/ImagesGallery";
+import ImagesGallery from "./Pages/ImagesGallery/ImagesGallery";
 
 function App() {
   const router = createBrowserRouter(
