+++ src/api/imageSearchApi.tsx
@@ -58,40 +58,9 @@ export const imageSearchApiCall = async (query: string, page = 1) => {
       author: image.user || "",
     }));
     const total = result.total || 0;
-    return { images, total };
+    return { images, total } as IImageSearchReturn;
   } catch (error) {
     console.log("error", error);
-    return { images: [], total: 0, error: true };
+    return { images: [], total: 0, error: true } as IImageSearchReturn;
   }
-
-  //   .then(async (response) => {
-
-  //   })
-  //   .catch((error) => {
-
-  //   });
-
-  // return response;
 };
-// export const imageSearchApiCall = async (query: string, page = 1) => {
-//   const response = await fetch(
-//     `${baseURL}?key=${APIKey}&q=${query}&per_page=10&page=${page}`
-//   )
-//     .then(async (response) => {
-//       const result: IImageResponseResult = await response.json();
-//       const images = result.hits.map((image: IImageResponseItem) => ({
-//         url: image.largeImageURL || "",
-//         id: image.id || -1,
-//         alt: image.tags || "",
-//         author: image.user || "",
-//       }));
-//       const total = result.total || 0;
-//       return { images, total };
-//     })
-//     .catch((error) => {
-//       console.log("error", error);
-//       return { images: [], total: 0, error: true };
-//     });
-
-//   return response;
-// };
